pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sn-ai-migrator"
version = "0.1.0"
description = "ServiceNow AI Migration Architect — analyze SN instance and generate migration blueprint"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "typer>=0.12",
    "rich>=13.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "openai>=1.0",
]
license = {text = "MIT"}
authors = [
    {name = "AI Transformation Architect"}
]

[project.scripts]
sn-ai-migrator = "src.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "black>=24.0",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.black]
line-length = 100
target-version = ["py311"]
